API session management.
Detect connection / disconnection of the client to the server API.
This allows to manage an APISession, associate objects with it,
and ensure that they are cleaned up when the session terminates.

This will be used to ensure shell session terminations on server
side, even if the client terminates unexpectedly (e.g. kill -9).

To do: make it work when nodes connect to NSAPI too.