Fix image shell session.
This makes heavy use of things developped in previous commits.

There is no server object accessed remotely from the client to manage
the shell session any more.
This object is now only visible on server side, and the client
manipulates this through regular API calls (create and save a session).