Handle unexpected client disconnection.
Typical case:
$ walt device rescan
^C
$

The server performs API calls to the client in order to
display messages on its std streams, or to get the walt
username.
Such a disconnection was causing those calls to fail on the
server side.
We now use a FailsafeRemoteAPI objects that ignores this
kind of exceptions.
When it occurs, the result of the API call is None.
The server sometimes has to check this case (e.g. when
it tries to retrieve the walt username).