xxx to server communication now uses plain sockets (no rpyc)
The client has to check the CSAPI version of the server.
Previously, it was doing that throuh the regular client to server
connection, based on RPyC.
This could cause reproducibility issues, if the rpyc library
is not maintained anymore in the future.
Thus we now use plain socket for this communication, and since
the mechanism is defined in walt.common.apilink, this also
changes the way nodes and the walt-dhcp-event program communicate
with the server.
The hub thread of the server now uses a TCP server mechanism
similar to the one of the main thread. The intent is that in the
future, all interactive services currently handled by the tcp
server of the main thread will be handled by the one of the hub
thread. For now, only the API service is handled by the hub thread.
Also note that, as a side effect, API communications between the
client and the server are faster now.