Implement Hub task asynchronous response mechanism.
This allows to fix the following:
* walt image search
* walt image clone
* walt image publish
* walt node wait
Previously, the client was sending a response queue as a parameter.
In the case of walt node wait, the response queue was saved until
the node bootup was detected. At this time a message was sent to the
response queue in order to unblock the client.
In the other cases mentioned above, the server was passing the task
to the blocking thread which was performing the long processing.
Then, the main thread was woken up again to push back the result to
the client.
With the new architecture involving the hub thread, it is no longer
possible to pass such an object (the response queue). Moreover, in
some sense it is no longer relevant, since this kind of wait on the
client side is already implemented for all operations.
The only issue is that the main server thread is expected to do the
processing and return the result to the hub thread. In this case,
this cannot be applied, because when the main thread returns, the result
is not ready yet (e.g. it is being computed by the blocking task).
Thus, this commit implements a mechanism allowing the main thread
to change a given task from 'sync-mode' to 'async-mode', in order
to indicate that the result will not be available yet when it
returns.