device tree: rewritten to print where the parent switch is connected
This involves a large update of topology.py and tree.py.
The trickiest part is to keep obsolete connectivity info, which
is very useful when debugging connectivity problems (cf. devices
appearing with parentheses), while still maintaining a coherent
network with no loops, even if the user moved a switch elsewhere
on the network.

The topology table of the database has changed.

walt device show will have to be adapted.