device registration: always update properties
Previously 'register_device' would only add new devices. This behavior
would not allow device properties to change, for example formerly
unknown device could become a switch after a software update.

This changeset removes 'add_if_missing' and changes 'add_or_update'
to add new devices or update their properties, in which case it
returns true, false is returned if the device exists and is unchanged.

'register_device' is almost useless, it only tries to set a static IP if
the device is a switch, then calls 'add_or_update'. This should be
refactored to create device classes and provide generic behavior in
'register' in the Device superclass and specific behavior in sublcasses
Node and Switch. Switch can be customized further to adapt to specific
manufacturers.