Optimize image creation date / top layer handling
Updating the creation date of a specific image is a heavy
operation. Since this was done for all images at startup,
it was causing the daemon to hang a long time a startup.

The creation time is linked to the top layer of the image:
if a container is committed on top, the date of the image
is modified.

Before, these values where updated at startup and after
each image modification.

Now:
* at startup the modification time is computed much more
  efficiently (we get the dates of all images at once
  by communicating only once with the docker daemon)
* the top layer id is computed on demand (cheap operation
  since it just have to read a json file), thus we do not
  have to call an update method each time we modify an
  image (clone, shell, etc.)
* when the modification time is requested (walt image show),
  we check if it was already computed, and if the top layer
  did not change since we computed it. Thus we recompute
  it only when needed.