Fix PoE reboot on stacked switches
In POWER-ETHERNET-MIB, PoE ports are identified using a
tuple (grp_index, port_index).
grp_index identifies a PoE port group (box in the stack,
module in a rack, etc. or 1 for non-modular devices).
port_index identifies the PoE port within this group.
On the other hand, LLDP ports are identified using a global
index meaningful for the whole switch.
When we want to alter PoE state on a port, we have to compute
the appropriate (grp_index, port_index) given the LLDP port index.

Previously, we did only comply with non-modular switch devices,
thus we used (1, <lldp_index>) to target the appropriate PoE port.

In order to comply with modular switch devices, I have found
the following solution:
- list the switch ports with speed 10Mb/s 100Mb/s or 1Gb/s
  (these are the possible speeds for a PoE port), using IF-MIB
- list the PoE ports, using POWER-ETHERNET-MIB
- verify that these 2 lists have the same length
- "zip" these 2 lists in order to attach each LLDP port to its
  corresponding PoE port identification tuple.

IMPORTANT: This only works with switches where all ethernet
ports are PoE capable. If this is not the case (the 2 lists do
not have the same length), WalT throws an exception.