Rework RPC communication between server threads.
We were using the rpyc library between the hub thread
and the main thread.
A bug was sometimes occuring in this code (in the hub thread).
It might be due to concurrent accesses to an object of the
hub thread (maybe the "requester" object, that was sometimes
also accessed from the 'blocking' thread).
Bug occurences were rarely appearing and I did not manage
to exactly understand why it was occuring. Thus I decided
to change this part of the code and avoid using rpyc
(I doubt the bug was in this library, but we have little
understanding of it since we did not write this code, which made
the bug hard to solve). It was the last part of the code that
was still using rpyc.
The dependency to rpyc is now completely removed.