Fix walt [image|node] cp when using minimal images
In order to preserve file attributes and allow to
transfer both files and directories, tar processes
are used at both ends of the transfer.
We the file or dir transfered is renamed, we were
using a tar option --transform that is not available
with busybox tar.
We now use a temporary name during the transfer
(because using the source or dest name could cause
file overwrites if the same filename exists on the
other end).
On the sender side, we first symlink the source
using this name, then tar this symlink (using the
tar option to follow symlinks). On the receiver
side, we untar and then rename the file to the
final dest name.
There was also an issue when using docker with option
'-i' in a popen object, on sender side: the docker
run command was never ending.
Since this option is not useful on this side, we
removed it.