Rework node's ssh host authentication (sshd + dropbear)
We no longer use the '-q' switch of ssh. This allows to let the user
know what happens in case of ssh connection issues (e.g. ssh server
not ready yet after a node has rebooted).
Host fingerprinting was not working correctly with dropbear-based
images because dropbear does not use the same directory and format
as sshd to store its host key.
The server code now copies the appropriate files to handle those
2 possible services when mounting an image.
The host key is the same but encoded differently for sshd and dropbear.