walt node deploy -> walt node boot (for clarity)
WalT images are never really "deployed" on nodes, in the commonly
understood sense of that term: the operating system is never transfered
and written on nodes.
Instead, nodes use a network boot procedure, and access the operating
system files on the server, on-demand, using the Network File System (NFS).
To clarify this, we renamed 'walt node deploy <nodes> <image>' to
'walt node boot <nodes> <image>' and rephrased appropriately the inline help
and messages printed.
For backward compatibility, 'walt node deploy' is still available, but it
is marked as an 'alias to the "boot" subcommand' in the online help.