Fix walt image search (really, this time)
The docker search command and corresponding docker-py function
returns only a small set of results: default is 25, and it can
be increased up to 100.
This is why some of the walt images where not found when typing
'walt image search'.
We want all available results. The only way we found to achieve
this is to interact directly with the docker hub json API, which
returns paginated results (max 100 per page, again).
This commit implements this.