server: add support for netconsole
Netconsole is a Linux kernel module that can send kernel logs over the network.
This can be very useful to troubleshoot low-level issues on nodes.

This commit adds support on the Walt server to collect these logs and store them
in the existing logging facility.

Example usage:

$ walt log show --realtime
13:31:48.280624 lofoten5.netconsole -> [   11.068686] systemd-journald[233]: Received request to flush runtime journal from PID 1
13:31:48.405583 lofoten5.netconsole -> [   11.195892] acpi_cpufreq: overriding BIOS provided _PSD data
13:31:48.425458 lofoten5.netconsole -> [   11.212867] input: Power Button as /devices/LNXSYSTM:00/LNXSYBUS:00/PNP0C0C:00/input/input0
13:31:48.429580 lofoten5.netconsole -> [   11.222108] ACPI: Power Button [PWRB]

To benefit from this, nodes should run a kernel with netconsole support,
either built-in or as a module:

    CONFIG_NETCONSOLE=y
    CONFIG_NETCONSOLE_DYNAMIC=y

Also, netconsole is configured through the kernel command line.  This is
automatically done by Walt for the pc-x86-64 node type.

Netconsole documentation: https://www.kernel.org/doc/Documentation/networking/netconsole.txt

See TODO in the code for some remaining issues to fix.