ipxe: reboot if kernel fails to load
On some nodes, for instance the PC Engines APU2, the bootloader launches
a memtest when iPXE fails.  This is annoying because it requires a hard
reboot (power cycle) for the device to try again to boot a Walt image.

So, make sure iPXE reboots the device when it fails to load the kernel.