ipxe: add device "submodel" to allow images to override kernel or kernel parameters
In addition to the current "/<model>/" directory, we now allow Walt images to
place a kernel, an initrd and a custom iPXE script in "/<model>/<manufacturer>/<product>/",
where:

- <model> is the current Walt model, for instance "pc-x86-64"
- <manufacturer> is extracted from the BIOS, for instance "PC Engines"
- <product> is also extracted from the BIOS, for instance "apu2"

This allows to accommodate for the wide variety of x86 devices: a single
walt image can now work optimally on a range of devices, customizing things
like the serial console or the netconsole interface name depending on the exact
device.  It is even possible to use a custom kernel for a particular device.

The iPXE boot script tries to load first "/pc-x86-64/PC Engines/apu2/kernel", and
falls back to "/pc-x86-64/kernel" (this example is for a PC Engines APU2).
The initrd is loaded in the same way.

For the custom iPXE script, *both* model and submodel script are loaded,
in this order.  This allows a submodel iPXE script to override variables from
the generic iPXE script for the model.  For the PCEngines APU2 for instance,
we would load:

1) /pc-x86-64/boot-params.ipxe
2) /pc-x86-64/PC Engines/apu2/boot-params.ipxe

To find out what is the "manufacturer" and "product" for your device, look at
the boot messages from iPXE, after "Walt node submodel".