Fix walt node cp file rights when node/image is target.
walt node cp transmits files or directories using tar
archiving on sending side and tar extracting on receiving
side.
The default behavior of tar is not the same when user using
it is root or not.
When extracting as user root, file ownership is preserved.
When extracting as a regular user, file ownership is set to
this current user.
Thus, when copying a client file or directory to a node or image,
ownership was preserved, since the user is root on the receiver
side.
This creates files belonging to a user whose numeric ID is
unknown on the node or image.
This commit adds a chown command after the tar unarchiving step
in order to solve that.