walt device tree: verify that client can print special chars
The client may type:
$ walt device tree > tree.txt
or
$ walt device tree | grep -C 10 "my-switch"
In this case the walt command should send ASCII chars only.

Moreover, if walt client is installed on a machine with limited
encoding, printing special characters may cause UnicodeEncodeErrors.

We now retrieve client's stdout encoding, and verify that this
encoding allows to print the special chars we use for the tree.
If not, we fallback to a set of ASCII chars.