Fix tree root selection
We selected the first neighbor of the walt server as tree root; however,
from the implementation of virtual nodes, we may select them as tree
root, that makes `device tree` displays only this one.

Now, we select the first neighbor of the walt server *which is a switch*
as tree root.