New image naming (walt_name = <repo>[:<tag>])
Previously, the image name in walt was the docker tag.
That was not very handy regarding scientific publications
because there is no tag-specific description page we can
link to on the docker hub.

This commit implements a more natural mapping. The name
in walt is now the name in docker (plus, optionally,
a tag).

In order to implement efficient search of walt images,
we now have to maintain a special image at
<user>/walt_metadata:latest.

* The existence of this image indicates that this docker
  hub account contains walt images. And we can use
  "walt_metadata" as a keyword to retrieve all walt users.
* In this image, we store a LABEL (cf. dockerfile command
  with the same name) metadata = <base64-metadata>
  Within this metadata, we have the list of walt images
  in this docker account, and, for each of these images,
  a cache of their labels (because it would be expensive
  to retrieve them one by one).
* The "walt_metadata" image is updated each time the user
  uses 'walt image publish'.
* It is also possible to force this update by using
  command 'walt advanced update-hub-meta'. This could be used
  when docker images have been built and pushed to the hub
  without using walt. This command will scan the images
  on the docker hub account of the user. We know if a docker
  image is a walt image by checking if it has a LABEL
  'walt.node.models' (see below).

All walt images must provide a LABEL called
'walt.node.models' that indicates the node models the image
is compatible with (e.g. walt.node.models="rpi-b,rpi-b-plus").
Note that this commit does not include any management about
this compatibility information. But it should be implemented
shortly.