Fix interactive.py when used with an input pipe
This allows the following kind of commands to work:
$ echo '\dt' | walt advanced sql

Before, it was not working because the walt command
stopped as soon as its stdin was closed, thus the
server response was not obtained in time.

Now, a socket.shutdown(WR) syscall is performed
to let the server know its input stream is ending.