Image mount path now based on image id.
Before, an image having several tags could be mounted several
times (once for each tag). This was probably the cause for
some issues regarding NFS mount refreshing (issue 33).
Now, the key of image mount path is the image id, so an
image can only be mounted once.
As a side effect, we now have a distinction between "mounted"
and "in-use". An image might actually be mounted because
one of its tags is "in-use", even if the other tags are not
used by any node. Therefore, we modified the checks in various
files: mounted -> in_use.

fixes: #33