topology: fix corner cases.
Displaying a tree of walt devices as shown by walt device tree is
actually a complex process.

The data comes from last walt device rescan command. However:
- Some temporary failures may occur (a switch may not respond
  to our SNMP request because of a network problem, for instance).
- Devices may be unplugged or moved (switches and nodes).
- When disconnecting a device from a switch, the switch will still
  include this neighbor in its SNMP responses for a few minutes.

In the database, the last mac address seen on a given switch port
is recorded, together with a flag called confirmed indicating if
this mac address was still there at the last scan or not.
Then, a complex process has to analyse this data. Since devices
might have moved (or because of point 3 above), some loops may be
seen in the data. This process has to solve them, by discarding one
of the links in the loop. It will preferably discard a link with
confirmed=false, but sometimes all links of the loop are recorded
with confirmed=true (because of point 3). In this case, it was
discarding a random link of the loop.

Two improvements are implemented in this commit:
- a bug fix: recently, walt device rescan command was modified to
  accept an optional argument to specify which switches should be
  queried. But, in all cases, when this command is run, all links
  in the database were re-initialized with confirmed=false before
  the scan was performed. Obviously, this re-initialization should
  be applied only to the links related to the switches which will
  be queried. This commit implements this fix.
- an enhancement to the discovery process: it now takes into account
  the fact that a node can have only one neighbor in the topology,
  which greatly improves handling of cases 2 and 3 above.

This fixes issue #45.