Fix walt image shell hanging (podman+systemd bug)
walt image shell was hanging when walt-server-daemon is started as a
systemd service.
When running walt-server-daemon manually or in a screen session for
instance, there was no issue.
This is due to a bug of podman related to systemd notify management
(walt-server.service unit file as a type=notify). podman container
creation was hanging.
This commit fixes this by clearing the 'NOTIFY_SOCKET' env variable
just after walt main process has sent its "READY=1" notification to
systemd.