Improve nbfs integration.
Sending SIGHUP to the daemon now makes it reload the server spec file.
This allows to change the set of exposed features without restarting
the daemon.
However, note that an image has to be un-mounted & remounted before
a change in the features can be taken into account.

The server now always updates nbfs exports file (when triggered by
a change in image mounts) if the nbfs feature is declared in the spec
file. Before it was only if nbfsd PID was found. Then a SIGHUP is
sent to nbfsd daemon if possible.