Move fake ipxe feature to walt.common package
For now this fake ipxe bootloader feature is used in the case
of virtual and vpn nodes.
In both of these cases it was part of walt-virtual-node code.

In order to reuse this feature for kexec reboots, we will
have to call this feature from walt.node package.

In order to prepare this change, this commit clarifies the limit
between the code related to this fake ipxe feature and the code
related to walt-virtual-node. And the code for this feature
in moved to file walt/common/fakeipxe.py.

Next steps:
* handle HTTP requests for more performance
* reuse the code for kexec reboots