server: implement grace time before image umount
When the last node using an image is associated to another image,
this previous image should be unmounted since it is not used anymore.
however, if this image defines /bin/walt-reboot, rebooting the node
involves code from this previous image. In order to allow this pattern,
we implement a grace period before an unused image is really unmounted.