fake-ipxe: allow ${<undefined-name>} (caution, see commit long desc)
iPXE allows referencing an undefined variable, using ${<undefined-name>}.
In this case, no error occurs, but the reference is just replaced by
an empty string (this is like unix shell scripts).

Previously we did not handle this case in our fake ipxe bootloader,
thus an error was fired.
We know act the same as iPXE.

However, previous versions of fake ipxe code may still be present in
walt images or VPN node images; thus *for now and in the near future
this kind of code should be avoided in all ipxe boot scripts*.

Simply defining an empty variable remains compatible with older code:
```
set emty_var
[...]
... ${empty_var} ...
```