new user: clone default images in more cases
The previous behavior was not logical.
If first command of a new user was, for instance,
"walt image shell pc-x86-64-default", an error
was printed because the user did not own any image yet.
But if second command was
"walt image show", then image "pc-x86-64-default" was
possibly cloned and listed.

When looking for an image by its name, if this requested image
is missing and the requester has no image yet, server now
runs the procedure to clone the defaut images too.