Rework systemd services definition
Add walt-server-netconfig.service: this service brings the walt
network up/down. Other WalT services may depend on it.

Add walt-server-dhcpd.service: this service handles the DHCP server
dedicated to the WalT network. It is turned on when the network is
on. When walt-server-daemon is up and is updating configuration,
it reload-or-restart this daemon to notify the configuration change.

Update walt-server-httpd.service: now, walt-server-httpd listens on
walt-net IP address only. This allows other http services to possibly
run on other addresses.

By default, walt-server-httpd.service is enabled and WantedBy
walt-server.service, so that it automatically starts whenever
walt-server starts. But walt-server.service still has to be enabled
to auto-start on boot.

Add nfsd service name in server.conf. Service name may depend on the
distro or on the administrator's configuration.
/etc/walt/server.conf now supports configuring these names.
Added doc on that topic.

This commit also allows distro packagers to specify systemd units
installation prefix when running walt-[server|node]-setup.
The default remains /etc/systemd/system, so there is no change for
usages not aware of this new option.
walt.*.setup modules have been reworked to have:
- a `__main__` module, executable with minimal code
- a `__init__` module to handle command-line interface
- dedicated modules for real tasks (systemd install, busybox install…)

Fix #63