Drop default shell=True when 'do'-ing commands
shell=True implies a unnecessary shell start and command parsing.  We
replace that with `shlex.split` to transform the str to a list of args,
then just fire the command without shell.

The `do` function has now a parameter `shell: bool` which allows to
still request for a shell if this is the intended operation.