Activate all interfaces when IFACE is unset
User may ignore exact interface name to activate; in which case, it
won't specify an interface and the script will activate all of them.
Note that most of the time, this is what we want when launching the
script by hand.

ifupdown is still compatible, as we still support IFACE variable.