vnodes: use vhost=on for better network performance.
Vhost-net allows to offload virtual network adapter work as a
kernel thread on the host, allowing better performance.

This option is only available when network device is specified with
type=tap. Before, we were using type=bridge.
So now we have to handle the linking of the tap to the walt-net bridge
ouselves. This is done by scripts walt-vnode-if[up|down] specified
on qemu command line.