vpn: use l2tp interfaces (naive implementation)
TAP interfaces are slow because we can only read and write packets
one at a time. This commit allows to establish the VPN using L2TP
interfaces instead of TAP interfaces. This will allow to manage
several packets at once using sendmmsg() and recvmmsg() in follow-up
commits. For now the implementation much ressembles the previous
TAP-based implementation.

Since we transfer packets through a SSH connection, we obviously do
*not* create  a direct L2TP tunnel between the client and the server.
Instead, on each side (client and server), we create an L2TP tunnel
on the loopback interface. The kernel believes it is communicating with
the remote peer, whilst it is actually communicating with a userspace
component (walt-vpn-client and walt-vpn-server, respectively). This
component transmits and receives L2TP packets over the SSH connexion.

Check the long comment in vpn/walt/vpn/server.py for more info.