walt-vpn: improve scalability & security
walt-vpn-server now just creates a tap and sends its file descriptor
to walt-vpn-endpoint through socket message ancilliary data.
Next walt-vpn-endpoint manages the transfers between the tap and the
ssh channel.
This is more scalable than letting the server manage all clients alone.
Moreover, walt-vpn-endpoint and walt-vpn-client now share most of the
same C code for these transfers between the tap and the ssh channel.

On the security front, the VPN Unix socket is now only readable /
writable by the walt-vpn user. Previously, this user was created
when walt-server-daemon was first run.
This implies that:
* walt-vpn-server must start after this user has been created
* walt-vpn-server.socket systemd unit cannot be started before
  this user exists.
In order to deal with this, creation of walt-vpn user was moved
to walt-vpn-setup tasks.

Fixes #70.