get_user_image_from_name() fix: disallow image fullname as input
For the "walt node cp ... booted-image" feature, function
get_user_image_from_name() had been modified to allow specifying
an image fullname instead of a user image name as input.
This was not a good idea and caused issues (cf. issue #72).

This commit reverts this change and handles the "booted-image" case
before calling this function.

Fix #72.