walt [node|image] cp: introduce walt-tar-send to fix rare tar segfault
File transfers are implemented using a combination of "tar c" (sender)
and "tar x" (receiver).
Sometimes, the file or directory transfered must also be renamed:
$ walt node cp <node>:/root ./node-root-dir
Before this commit, we relied on a trick on the sender side,
creating a symlink to the source in a temporary directory, and using
tar "-h" (dereference) option to have the tar archive member renamed
as expected.
Unfortunately, in the case of a directory, tar "-h" option applies
to the whole sub-tree which may cause tar segfaults in some cases
(such as infinite symlink resursion?).

This commit removes the use of option "-h". If the archive member
does need a rename, a copy in a temporary directory is done before
sending. If not, "tar c" command is started right away (fast path).

Since this process starts to be complex, a script `walt-tar-send`
is introduced. It is copied by the server when an image is mounted,
thus it can be found on all nodes. For copying from an image (to the
client), an option is added to the relevant podman command in order
to use the same script:
"-v <walt-tar-send-path-in-source>:/bin/walt-tar-send"