Generate internal platform logs
This commit lets the server software generate walt logs
when important events occur.

Currently, logs are generated:
1. when a new device is detected after its DHCP request.
2. for saving walt-server-daemon stdin and stdout.

About 1: the documentation was updated to reflect usage
of these logs. When a user connects a new node it may
run the following to verify it is properly detected and
to get its name:
$ walt log show --platform --history -5m: "new node"
A similar command is proposed when connecting a new switch.

About 2: systemd is sometimes not very reliable regarding
logs. Having stdout and stderr saved as log lines in the
database allows to deal with this. One can type for instance:
$ walt log show --server --history -1h: > server.log

Note that walt node [show|wait] commands have been modified
to handle these new features.
* option "--nodes" has been renamed to "--emitters" since
  the server may now generate logs too.
* a new "--platform" option is introduced as a shortcut for:
  "--emitters server --streams platform.*"
* a new "--server" option is introduced as a shortcut for:
  "--emitters server --streams daemon.*"

Fixes #52.