Update node bootup notification system to detect changes
On bootup, after local ssh port 22 is opened, nodes now open a
permanent TCP connection to walt server.

When this connection is detected by the server, node's booted
status is set to true. If the connection is closed or reaches
a keepalive timeout, the node's booted status is set to false.

On node side, if the connection closes, this means the server
is down. In this case, the node reboots too.

When the booted status of a node changes, a platform log line
is now emitted.

Fixes #25.