New command 'walt node console'.
Allows to display a serial console for virtual nodes.
It is based on qemu emulated serial console. Since output of qemu
process is now saved as walt logs, this command is based on the
retrieval of these logs in realtime.

The input of the console relies on a new CSAPI function which
writes the input bytes on the stdin of the qemu process.

We do not use the alternate screen buffer of the terminal because
this would discard the terminal scrolling feature.