walt-server-httpd: communicate with walt server through unix sock
walt-server-daemon now listens on a unix socket to serve the needs
of companion daemons such as walt-server-httpd.
The first request implemented is REQ_FAKE_TFTP_GET_FD, allowing
to open a file on behalf of walt-server-httpd and transmit the
open file descriptor using unix socket ancilliary data.
This avoids having to transfer the whole file content through
the socket, thus improves performance, but it keeps the
responsability of locating node files to walt-server-daemon.