server dhcp event: fix unknown device -> node transition
Users sometimes connect new nodes to the platform without
a proper bootloader configured. This is useful to check
that the hardware is working for instance, and/or when
developing the support of a new node model.
With the current code, walt-dhcp-event was no longer
called after the first DHCP request (because the device
already had a host entry at subsequent requests), thus
the main daemon could not detect that the dhcp UCI or VCI
option was now properly set to 'walt.node.<something>'.

This commit changes the DHCP configuration template
in order to start walt-dhcp-event whenever the device type
is unknown.

However, the code of walt-dhcp-event was also modified to
filter requests more finely and avoid bothering the walt
server main daemon with useless requests.