node registration: simplify handling of default images
This commit removes the "ready" column of db table "images",
and the 'ready' attribute of "NodeImage" python objects.
The process no longer needs the concept of images 'not ready'.

This concept was introduced to show that a default image was being
downloaded for a node. So the node was not readily usable, and
the image was not yet present in walt (podman) internal repository.

This concept introduced several bugs in the past, with corner
cases such as a failing download of the default image.

Now, the first time a node of a new model is detected, it is first
registered as a device of type 'unknown', and then turned into
a walt node only when the download of the default image has succeeded.

If the download fails instead, it remains a device of type 'unknown'.
When the device reboots, it will send a new DHCP request, which will
cause a new retry of image download; it this succeeds, it will finally
be converted into a real walt node.