evloop: discard recursion depth restriction for running events
This was causing subtle issues, in case of recursive loop calls
(e.g., because of a synchronous rpc call), and with some events
registered at a recursion depth > 0 whose callback was waited on
to decrement the recursion depth.
Moreover, the main process can very well be at a recursion depth
> 0 for a moment (without being blocked, that's the purpose of
event loop recursion). This could prevent urgent events to be
processed in time (such as node clock synchronization).