server: avoid defragmenting logs when unsynchronized
If getting wrong timestamps with received log lines,
the process of defragmenting logs can be very resource
intensive for the database.
We actually restrict processing of the queries involved
to the oldest timestamp we recently recieved, but if getting
for instance ts=7 (7 seconds since the epoch), we are
more or less considering the whole history of logs.
So it we detect that timestamps are not properly
synchronized with current time, just avoid defragmentation.