server: refactor startup of walt-server-daemon processes
We now use set_start_method("spawn") -- the default for python >=3.14
instead of "fork".
With "spawn", objects started by the main daemon process before
starting Process() objects are pickle / unpickled to be transmitted
to the new child. So some code refactoring was necessary to avoid
this early creation of complex objects that cannot be easily pickled.
The result is that for example server-main objects are now in their
vast majority created by the run() method of the child process,
and they are no longer present in the memory of sibling processes
server-db, server-hub and server-blocking.
So this commit is future-proof and it decreases RAM usage.
Some imports were also moved to the functions which really need them,
to further improve this RAM profile.

How we manage the end of subprocesses was clarified.
* daemon uses p.join(), p being the Process() object.
* sub-processes (such as p above) also start subprocesses, by using
  walt.server.popen.BetterPopen or other means. They use automatic
  os.waitpid() calls when they receive a SIGCHLD signal.

os.setpgrp() is now called both on the daemon process and
on its walt-server-daemon:server-[main|db|hub|blocking] subprocesses,
to ensure signals only reach the target process.

After these changes, the redirection of pdb input (when inserting
a pdb.set_trace() line for debugging) in subprocesses had to be
reworked. The subprocess now requests the daemon process to read
and return an input line.