rpi3b+ & rpi4b: remove the need for a SD-card on the 1st boot
All free IP addresses now have a link nodes/<free-ip>/tftp targeting
a directory of default boot files.
If the board does not end the DHCP handshake, walt-dhcp-event is not
called by dhcpd so walt-server-daemon is not aware of the new node
trying to boot. But with these TFTP links, the node can still find
the u-boot binary, which will take care of detecting the rpi model
and redoing a proper DHCP handshake. Then the server will be able to
prepare a default image and the node will be able to boot.

Fix #103.