wip -- analyse hot points of trackexec trace
We want to count how many times each couple (file_id, lineno) is found in the
trace.
With that we can easily find hot points in the program.

Current issue: numba does not know argument return_counts of
np.unique().
Try to move the 3 last lines of the numba function out in a
calling function (plain python function).
See if it does not slow down the thing too much.

Maybe we could add a different utility trackexec-hotpoints
to display this kind of results, instead of the viewer.

Note: we will have to rebase because we have earlier commits
tagged wip before this one.