server: optimize BetterPopen and image "filesystem" objects
Image "filesystem" objects allow to start a "podman run" process
on-demand for helping "walt image cp" command validation or
completion. Those processes are kept running and stopped when
they are no longer used for 1 minute.

Closing those processes involved a synchronous popen.close()
operation, which could take some time. This was slowing down
server-main, and very visible when the process ends and stops
several of those running "filesystem" processes.

Those filesystem processes now rely on the new script
walt-server-fs-helper, which adds a temporary name (tag)
to the image, and runs "podman run" on it. This prevents
problems if the image is renamed or removed and this
process is still running.

This commit removes several calls to image.filesystem.close()
because they are no longer needed: we can just let the
filesystems cache stop the process after the timeout is reached.

The BetterPopen object handling the subprocess was also
refactored to use pdfd_open() and better interact with
the event loop.