server: fix waitpid() race condition
The previous code was automatically calling os.waitpid()
on terminating child processes because of a signal handler
set on SIGCHLD.
This could cause issues when calling subprocess.run()
in exttools.py, because if the signal handler is faster,
os.waitpid() is already done and the internal code
of subprocess.run() can no longer retrieve the exit
status of the command. Because of that, a failing
command could pass undetected even if check=True was
specified.

The current code explicitely calls os.waitpid() when it is
possible / easy, and implements ev_loop.auto_waitpid(<pid>)
for other cases.
ev_loop.auto_waitpid(<pid>) internally works by using
os.pidfd_open(<pid>) to get a file descriptor and be able
to detect when the process ends.