walt device config expose: fix possibly blocking connect()
When using walt node config <node> expose=<n-port>:<s-port>,
and the local browser pointing to http://<walt-server>:<s-port>,
the main server thread could hang for long periods (TCP timeout)
when we were rebooting the node.

This is because we were using a blocking connect() call.

This commit introduces an object NonBlockingSocket for handling
a socket connection in a non-blocking way (for connect(), read(),
and write()).
This object is now used in expose.py to overcome this problem,
and also reused in netservice.py for the purpose of code
factorization.

The code in netservice.py was already non-blocking (at least for
connect()), thus it was a source of inspiration for the code of
NonBlockingSocket.