server: always update nbfsd conf
Before, we were checking if server.conf had feature "nbfs".
But we now want to manage the NBFS version compatibility between
images and server. So we specify for instance "nbfs1" for version 1,
both in image.spec and server.spec.
It's not a problem to always update the conf of nbfsd. If nbfsd
is not installed, the pid file will just be missing, so no
process will be notified, but that's all.